008 part 5: the build condition stops reading one file, and AC-14 goes green
The central claim was enforced over `RecknZkEscrow.sol` alone, but `settleWithProof`
obeys the struct `RecknVerdictVerifier.verifyVerdict` returns. A constant-address
branch spliced in front of the proof check there IS a resolver — and it passed every
one of the four existing checks. `no-keys.sh` gains check 5 over that file.

Check 5 closes the file by six PROPERTIES — what it is permitted to contain — not by
a list of forbidden constructs, because a hole in an enforcement script is never
closed by adding the name of the construct that exploited it. Every pinned value is
transcribed from the spec (§6.4), not generated from the file the script checks:

  5a  the region is literal      one quoted line, and it is the import, verbatim
  5b  vocabulary is closed       exactly 43 identifiers, equality in both directions
  5c  declarations are closed    by count, and the one function is verifyVerdict
  5d  the body is two statements verify first, decode second, in that order
  5e  assignment targets closed  five names, no field, index or member of them
  5f  the skeleton is closed     20 normalised pieces, in order — the only clause
                                 that sees an ORDER or a VALUE

Sandbox controls, six mutants, no repository file written (clean copy exits 0 first):

  a tx.origin == 0x…1337 branch          -> 5b and 5d
  the verifyProof call deleted           -> 5b and 5d   (what a denylist misses)
  minDelta <-> maxDelta permuted         -> 5f, and 5f ONLY
  uint8 outcome moved to the head        -> 5f          (a real proof, false release)
  REPRODUCED = 1 / FAILED = 0            -> 5f
  constructor parameters swapped         -> 5f

The third is the one that matters: it adds no token, removes none, moves no count, no
statement and no assignment target, so 5a-5e are all silent on it. If check 5 were a
search for the constructs the other mutants use, that mutant would live.

Also in this commit, because the claim may not widen in the code and not in the
documents that state it (§9(2a)-(2c)): AGENTS.md §0 and CLAUDE.md's central-claim
block now name the second file, and no-keys.sh's own scope comment — which said "the
body of `contract RecknZkEscrow` only" — is corrected in the commit that made it false.

Two further pieces 008 owed:
- §3.4's four-token edit: `uint64` -> `uint256` for pre/post/minDelta/maxDelta in
  RecknVerdictVerifier.sol. The Rust side has committed uint256 since part 1, so a
  proof carrying a value above 2^64 would have made abi.decode revert on-chain.
- AC-11's half: seven `if (!vm.exists(F)) return;` gates became `require`. A missing
  fixture used to make forge green for a run that verified nothing.

Measured: forge 12 passed / 0 failed; no-keys.sh exit 0 with five sections;
docs-check.sh 9/9 stale claims absent, 11/11 replacements present, 0 tilde literals,
1/1 qualified ~34 s site, cycles.json matches 3/3 guests — AC-14 is green, and
AC-00, AC-00b and AC-14 all pass through ac008.sh.

+++ zk-verdict/contracts/src/RecknVerdictVerifier.sol
@@ -6,10 +6,10 @@ import {ISP1Verifier} from "@sp1-contracts/ISP1Verifier.sol";
 /// @notice Public values committed by the zk-verdict guest. Must match the
 ///         `sol!` struct in `zk-verdict/lib/src/lib.rs` (field order + types).
 struct VerdictPublicValues {
-    uint64 pre;
-    uint64 post;
-    uint64 minDelta;
-    uint64 maxDelta;
+    uint256 pre;
+    uint256 post;
+    uint256 minDelta;
+    uint256 maxDelta;
     uint8 outcome;
     bytes32 traceHash;
     /// @notice Commitment over the guest's prestate + predicate + plan. An escrow
